make include-path resolution using make optional on a per-project basis
Make can be invoked by the background parser if some include paths are
unresolved. However some projects do not use make for building and this
potentially induces an overhead for large projects. This patch adds a
project-wide option to disable this behavior.

This patch checks the setting thread safe when the IncludePathComputer
is created, which is always done in the main thread. An additional
assertion is put into place to make sure this continues to work as
intended.

REVIEW: 102883

Based on the patch by Alexandre Courbot, many thanks!