fix wrong assumptions in background parser
a) properly initialize wait duration for parse jobs,
which greatly improves the actual meaning of our unit test... turns
out we are not as bad as I've feared :) <300ms for 1000 jobs.
b) the order in which parse jobs are *run* is not defined, and
results in random failure of some unit tests. instead, assert on
the *creation* order, which is well-defined through the associated
priority