Support a simple filtering in the abstractfilemanager API
This is meant to be used for simple cases like VCS-directories or the .kdev4 directory. I didn't add a
default, since the list for that should be assembled by the team to make sure there's not too much in it
and also not too little. Its a rather efficient plain-string-matching using QSet for performance reasons.