prevent UI-lockup when adding lots of stdout lines to make output model
we now add batches of 50 lines and then return to the event loop,
continuing with the rest of the lines via a QueuedConnection

the unit test shows that (on my machine) the event loop is roughly
returning every 80ms. the test ensures that this time does not cross
the threshold of 200ms which is imo too much.

on a further note, it is interesting to see that optimizing the actual
adding of lines is not easily possible, the tons of regexp matching
is simply required... the only possible way would be to make the error
patterns configurable such that e.g. those for unused compilers or
unused toolchains are not queried.

BUG: 295361