Allow languages to specify whether reparsing of a file is required if whitespace is changed
Python requires files to be re-parsed on intendation changes, because
the indentation is part of the language syntax. This patch adds a
virtual function to ILanguageSupport which can be overwritten by language
plugins to specify whether require files to be re-parsed on whitespace changes.
The default implementation keeps the status quo, so all existing plugins
are unaffected.

REVIEW:104453