apply filter before executing quickopen
sometimes users are pretty sure about the result for a given
filter string and don't wait on the result to show up
but press return instantly. this used to be broken now
with the filter timer in place. with this patch though
we make sure the filter is applied properly before
executing quickopen and everything works like a charm