make extern scripts available through context menu
Now, the context menu in the project manager view, tab and editor
show a submenu "external scripts" which holds the available scripts
which can then be executed for the selected file(s).

scripts that depend on a user selection are only shown in
the editor context menu.

thanks to Ghislain Mary for the patch

REVIEW: 104133
BUG: 257703