fix quickopen filter-timer logic
- when the filter-change is cheap, i.e. the user just appended some
chars the the previous filter, then use the current number of items,
i.e. rowCount(). if that is below a threshold of 10k, directly apply
the filter
- otherwise check the total worst-case number of items, i.e. unfiltered
(requires API/ABI change to QuickOpenDataProviderBase). If that is
below the threshold, directly apply the filter
- in all other cases, use a filter-timeout of *300* ms which is also
the timeout in e.g. kfilterproxysearchline. The 150ms of before
where not enough, even when I typed quite fast the filter would be
applied inbetween resulting in sluggishness