properly restore project manager view on close/open of projects
Previously, whenever any project was opened, the state of the
view was restored, as it was when the session was last closed.

Now we restore the state per-project whenever a project is opened,
such that the above issue is resolved.

Furthermore we now also save the state whenever a project is closed,
which allows us to accidentally close a project, reopen it and have
the previous state reapplied. On closing a second project or the
whole KDevelop application, we forget all not-opened projects.