Fix usage of uuid
The code generating a string-rep of the uuid used toByteArray, while the
code actually using this string-rep required a QString. This can cause
problems when the encoding is not something ascii-compatible on the machine.

In addition toByteArray was only introduced in Qt4.8, so this change makes
KDevelop build against Qt4.7 again.