Improvements to the "Add Toolview" dialog.
Allow selecting multiple toolviews to be added in the dialog.
Allow adding a toolview by double-clicking on it in the dialog.
Disable toolviews that are already present in the current main window
area to prevent adding the same toolview multiple times.
Also disable toolviews that are already present in the current main
window area in the dock area context menu.

BUG: 165969
REVIEW: 104818