Fix generation of the config-file
With the #define x @x@ the generated config file would still always have
a #define x in it, even if @x@ is 0. This results in both ext-hashmap and
unordered-map to be enabled even if only the latter was found and hence
results in compile-errors.

Instead if one wants to base a define in a configure_file-generated header
on the value of a cmake variable #cmakedefine should be used.

This fixes building of kdevplatform on systems where unordered_map produces
a warning without the -std=c++0x compiler flag which in turn makes
check_include_file_cxx fail to find the header.