Don't use raw Declaration pointers across unlocked DUChain code paths.
Since the declarations might be randomly deleted by the cleanup thread
or something else, it is potentially dangerous to use raw declaration
pointers after the duchain was unlocked and locked again.

Instead, always use either the DeclarationPointer or an IndexedDeclaration.

Furthermore, it is often not a good idea to search for declarations
holding a read lock, then unlocking it, just to acquire a write lock
and finally do something useful with the found declarations. This
would only be a safe operation if we use e.g. DeclarationPointers.

Finally I severly doubt that the code snippets changed here have
a worse performance now. Granted, now there is no chance that another
thread can operate on the DUChain while we search for the declarations,
yet we also have less context switching and don't need the DeclarationPointer
indirection and thus might be done faster...

BUG: 302744