Enable keyboard shortcuts for "Find Uses"
I'm using this context-menu action quite a lot lately, so make me happy by
allowing to set a shortcut for it by adding it to the mainwindows
actioncollection. Needed to make it a KAction, for an unknown reason it
did not show up in the shortcuts list as a QAction.