Don't tell the user the make config has changed when he first sees it.
The reason it that we didn't set a default value, hence the default
which is probably 0 was taken. Yet our spin box has a minimum value
of 1. Thus, the KConfigDialogManager saw a property difference in
it's updateWidgets method, which is called from KCModule::load. Due
to that, it triggered a delayed call to changed(true).

Now we set a proper default of 2, which should be suitable for most
machines. This fixes the issue - yay!