Directly set the model's data when we toggle a checkbox.
By default the editor is only closed when it loses focus.
This actually means the 'apply' button *does* get enabled
properly, but the user first has to click somewhere in order
to loose the editor's focus.

Now, we just close the editor whenever a checkbox gets toggled
which should make the behavior much less confusing.

BUG: 304352