Protect access of QAbstractItemDelegate pointer using a QWeakPointer.
The standard output view allows plugins to provide delegates and
also allows them to keep the ownership of them. While powerful, this
seems to be buggy and can lead to crashes in some cases.

Now we should properly detect the cases where we try to access an
invalid item delegate and output a warning instead of crashing.

It would still be interesting to know which plugin is responsible
for this bug.

CCBUG: 207669