Let the OutputView take ownership of a job's model and delegate.
This is required to ensure proper functionality of the views
when e.g. a plugin gets unloaded. Previously, this would trigger
crashes due to access of invalid model or delegate pointers.

Now the view will stay functioning even if e.g. the MakeBuilder
gets plugin gets unloaded after a project was compiled, or even
if you close a project while it is being built.

It is still safe to access both, the model and the delegate,
from inside a job as long as the job is running. When a view
gets closed, the job gets cancelled anyways.

Sadly, this is an ABI break and thus cannot be backported to
the KDevplatform 1.4 branch...

BUG: 207669
BUG: 204399