Restore working prev/next output marker behaviour
This was broken during the refactoring when moving the code from kdevelop
into kdevplatform. The shuffling of code apparently led to the wrong
lists size being appended to the set. The original code appended the size
of the actual lines in the model (not the error-items) to the list, which
is the same as the row-number of the newly-inserted item. This was
transformed into adding the size of the error-items-list, which simply
makes no sense. The code using this list actually expects row-numbers for
the model in the set.