Rename the member to reflect what it stores
This member was originally named errorItems in the KDevelop code, since it
does not store all activatable items. Warnings are also activatable, but
this set was introduced to hold errors-only and thus make the prev/next
actions navigate between errors and ignore other activatable items.

Restore that name for clarifying the purpose.