Get rid of BuildItemBuilderJob in Project Manager View plugin.
The whole and only purpose of that class is to convert KDevelop::BuildItem to
KDevelop::ProjectBaseItem*.

Refactor ProjectManagerViewPlugin slightly in order to separate
targets collection and job creation into different functions, and
incorporate the mentioned conversion into one of the functions.
This allows us to remove that class completely.

REVIEW: 106459