Astyle formatter: add support for padding parenthesis headers
The patch adds an option to enable/disable padding parenthesis headers
in the astyle code formatter.

For improved Qt compatibility, "foreach" is added to the list of headers
and it is assured that "&" and "*" are not treated as operators at this point.

The KDELibs and Qt styles are fixed to properly use this new feature.

BUG: 302593
REVIEW: 106032