Add headers to the MultiLevelListView.
This is quite a refactoring, but now the MultiLevelListView can
cope with arbitrarily deep nested models. The very last view
(i.e. to the right) will now show all remaining levels. Internally,
we now use QTreeViews instead of list views (since the former supports
headers, while the latter cannot handle them...).

Many thanks to steveire for writing the KSelectionProxyModel,
as well as giving me a rundown on how to use it.