Rework the UI for path-specific language params and split it off
This re-arranges the UI for the configuration of paths and the
includes/defines a bit to make it easier to understand and use. In addition
the code is split into a separate widget class since it doesn't really have
much to do with the rest of the configwidget and I felt the class got a bit
too big and hard to understand.

The widget is not yet fully functional, only adding/replacing paths
works at the moment.