Moved ConfigureAndBuildJob to KDevPlatform/utils
And rename it to SequentiallyRunJobs. It's a useful that lets you have a
list of jobs that should be executed sequentially.
I've moved it to KDevPlatform because it will be used from the cmakebuilder
and the makebuilder. Also it's very generic (only depends on KJob and
KCompositeJob) so maybe it's useful somewhere else, some day.