Add Thread and Frame Information on Execution
Moved the code that adds frame and thread information when not present
in the command already from the queueCmd method to the executeCmd
method and added some info at the debug strings to make clear that the
queued command will be modified upon execution. This is needed in order
to be able to queue commands whose information depend on previous
commands in the same queue.

CCBUG: 301287
REVIEW: 105210