Move plain-Qt application templates into a Qt category
The categories should indicate what kind of template it is, not where the
templates come from (contributed by a KDE or Qt or other third-party
project). So the CMake/Qt templates should be in a Qt category.

This also helps finding them for people that don't know the category system.