Revert "Properly identify faulted executions in the CommandExecutor"
This reverts commit e8710e1e552499fd76dad7d35737993619d458bd.

The reasoning is that a non-zero error code is not an error in executing
the process. The process did execute properly - unlike a crash or a
failure to start up at all.

In particular this change causes all users of the API to suddenly show
message boxes instead of simply outputing the error-code.