Fix a regression of AliasDeclaration handling in DUContext::findLocalDeclarationsInternal.
This was introduced by 764d457b: The ! operator takes precedence
over a & operator. Thus ! flags & Flag is the same as (!flags) & Flag
instead of !(flags & Flag).

The regression was only noticeable in PHP, since there we often search
for declarations by passing the DontSeachInParent. flag. If that is negated,
we get a SearchFlags that has the DontResolveAliases flag set, which would
then break the handling of closure uses.