Restore support for projects with only 1 builder.
This probably affects custom makefile, definetly affects custom buildsystem
and any other buildsystem plugins which have only a single builder. The
top-builder was not added to the list of kcm-names to be shown and hence
the project config dialog didn't show any configuration pages.

There should probably be a warning or even user-visible error shown when
the list of kcm-plugins to load into the dialog is empty since thats
always a strong indicator of a bug somewhere.