Ensure we do not listen to deleted signals for items in skipped folders.
Apparently, stopWatcher does not prevent the signal emission for deleted
files in a folder. Thus we emulate that in our own code and indeed that
fixes the issue that we could easily crash whenever a folder was deleted.

Thanks to Niko for the unit test and idea for the patch.

REVIEW: 10700