Remove the OpenWithConfig dialog again and instead make the menu usable directly.
Sorry for wasting your time, but we think that this dialog is not neccessary
after all. Instead we now show 'Default Editor' instead of 'Embedded Advanced
Text Editor'. That, along with the new sorting and grouping, should make
it straight forward for users to set the proper open with setting.

CCMAIL: Przemek Czekaj <xcojack@gmail.com>