Do not store const& to a shared pointer of AbstractType.
First: That doesn't make sense.
Second: It might lead to crashes if we assign temporaries.
Third: It doesn't save anything besides a potential refcount increase
in the shared pointer which we do so often its not worth optimizing
for in such potentially dangerous places.

Thanks to Maciej Cencora for the patch and hint.

REVIEW: 107035