Fix build
Fixes 9e14eee8433961f87df094c6a0fc11ed73b5e447 which adds a new target to
link to for the tests but does not define that target. The libs to link the
library against are copied from the unit-test, might be overlinked but I
don't intend to waste time on this since its not installed anyway.