Optimize: Re-use implictly shared project name and url.
The serialization to IndexedString does not save any memory,
actually the opposite is happening: We are breaking the implicit
sharing chain and thus need more memory once we want to display the
project name or need to access the project URL for displaying.

The sorting of the project files takes a very small performance
since we are now no longer comparing random string indices but
use the project file name instead. This should be nicer anyways.

Loading a project otoh is speeded up by a factor of ~2, according
to the benchmarks, e.g.:

before:
RESULT : QuickOpenBench::benchProjectFileFilter_addRemoveProject():"5000-f/b":
     42 msecs per iteration (total: 85, iterations: 2)

after:
RESULT : QuickOpenBench::benchProjectFileFilter_addRemoveProject():"5000-f/b":
     21 msecs per iteration (total: 85, iterations: 4)

Note though: For performance reasons we use QUrl and QIcon as
storage types. KUrl and KIcon have a broken implementation in
KDE4 with their additional dptr, something that is not required at all.
Just keep in mind that if you want to get a string out of the url,
use KUrl(theQUrl)... which is fast and gives you the non-broken
API. For KIcon you only need that for the icon lookup, i.e. at
construction not for saving it for later use.

The above was discuseed with dfaure. Imo we should adapt other
data storage points of KUrl/QIcon as well, most notably the KUrl
in the ProjectBaseItems.