Show empty edit field when adding define
Removing the "Double click here..." message upon editing makes it less
error prone to abort editing and clearer that the user can just start
typing.

Its also more in line with what line-edits do that show such a helpful text.

Thanks to Ivan Shapovalov <intelfx100@gmail.com> for providing the patch,
needed to redo it based upon the move to kdevelop and the last reworks in
the codebase.

REVIEW: 105693