Add a document list context menu button right to main tab bar.
Thanks to André Stein for the initial patch!

Original message:

This request adds a little button to the right of the tab bar.
It activates a context menu which shows (in sorted order) all
currently open documents and opens the selected document upon click.
This is a feature I miss from Visual Studio. I do know that it's
redundant to the Documents view panel but I think it allows a quicker
 access to files - especially when a lot files are open in the
IDE. As it is very small button I think it does no harm to users
who don't want to use it.

Adaptions:

- added an icon to the tool button, otherwise it just looked odd
- simplified the code (no STL e.g.)

Open issues:

- push this code somehow into shell or similar, to be able to share
code with the document switcher plugin. Esp. we want to access the
project model to get nice, short paths and icons.

- add the document status icon, but that needs to be updated when the
document is changed

REVIEW: 107170