Make the assistant behave more like a tooltip.
a) Do not impose the parent widget width as strict max width limit.
Instead we can now be as wide as neccessary, potentially spanning
over toolviews on the side and such if required.

b) Add the ToolTip window flag, which gives us a nicer look and
some hide/show animations for free.

c) Move the tooltip on resize events of the parent widget, such
that we always are where we want to be.

d) Move the tooltip *below* the parent widget. In KDevelop we usually
have a toolview border at the bottom, the perfect place imo to
show the popup.

TODO: handle the case where no bottom toolview border is shown.