Add status icons to document list tab bar button.
The status icon of each document in the document list next the tab
bar status label is now shown and updated accordingly. Additionally
the context menu is not built up every time but only if a new
document is added (was updated before with each action: remove,
title change and addition).

REVIEW: 107326