Move assistant popup just below the editor view. Also allow it to grow horizontally.
Now, the main window geometry is taken as a boundary limit, which should allow
for wider assistants even when we have any toolviews shown on the sides. Furthermore
we now never interfer with the editor, esp. not if one edits the last visible line
at the bottom of the editor.

Note also that kdevelop/sublime always reserves space at the bottom, either for tool-
views or for the progress bar. Imo this is a very good place to put the assistant.