Unbreak Open With plugin.
This regression was introduced with the ability to add custom actions
from other plugins. To do that we query all plugins for extensions
to the OpenWith context. This way, we query ourselves as well, which
then cleared our data structures, esp. the action map. This in turn
resulted in us never getting notified when an action was triggered.