Delay initialization of TextEditorWidget
This patch reduces the initialization of the loading a text document view
by delaying the initialization.
This is useful because when we change areas we close all the opened views
and then we reopen the new ones. The loading of document views was one of
the most expensive calls that happened there. This is now delayed to the
event loop so the view creation now is trivial by creating only an empty
widget that will be instantiated afterwards.

I'm not closing the bug for the moment, but it would be interesting to know
how it went, so feedback is welcome!

CCBUG: 291052