Prepare restructuring of file template tree.
The category is a user visible string and should be translatable.
Instead we now introduce two new config keys, Type and Language.

Type is optional and can be Class or Test, in which case some
additional config pages will be added to the wizard.

Language is also optional and references the X-KDevelop-Language
of a language plugin. This is looked up and then asked for a custom
createClassHelper.