If left or right toolbars are empty, don't show them
Otherwise we get some empty widget (about 5px wide) that is ugly and
annoying.
Hiding the toolbar is good because without it we get to have the central
widget in the right margin, so for example if we move the cursor in the
rightmost side in the screen, the cursor will be on top of the scrollbar,
and that is useful
Of course, the advantages require not having toolbars on the sides,
otherwise it won't be noticed.