Don't resize the vertical IdealButtonBarWidget when it's empty
This greatly reduces the flickering on area switching. What was happening
is that when changing areas all buttons were removed and re-added. In this
process the button bar would shrink when empty and that would trigger a
re-layout of the whole GUI. Now we keep the width of the buttons we've
added.

This won't be a regression given that my previous commit improves the
behavior when there are no buttons on the button bar, so we won't get
a big gray area.

Now all the flickering when changing area should be gone.

BUG: 291052