Implement OptimizedUrl::isLocalFile and support file:// URLs properly.
We now use KUrl as a url parser and store the url prefix separately.
This should allow us to support any kind of URLs as long as they do
not contain sub urls, or fragments or any other more esoteric feature.
These should not be used in file managers anyways imo, thus this
should be OK I hope.