Also add a functionality to get an IndexedString for an OptimizedUrl.
Note that this shows that the password handling in OptimizedUrl is
a problem. It should not be part of pathOrUrl but should be part of
toUrl if we want to keep 1 to 1 compatibility to KUrl...

For now we remove it and just ignore it... It would be lost anyways
when storing it an IndexedString but of course might be a serious
issue when opening files from a remote server. Though I think KDE
offers a way to store the password separately which should be much
safer anyways.