Use KDevelop::URL in the ProjectModel.
This is done in a way that requires minimal changes for now, just
to see how it works and how it performs.

The new API is optional now, you can still use the old KUrl way
which will now be even more expensive as before due to the additional
conversion.

Also, we are now more picky on URLs, as shown by the changes required
to the projectmodeltest: the rename test used relative URLs which will
not work with this change. I've ported it to the new API to fix it.

Furthermore, we use the new API in the projectmodelperformancetest now
to show the potential performance benefits. And indeed that is looking
very promising:

After:
init model 321
addBigTree 317
addBigTree 324
addBigTree 311
addBigTree 332
addBigTree 314

Memory increases about 32MB per addBigTree call

Before:
init model 623
addBigTree 608
addBigTree 616
addBigTree 601
addBigTree 628
addBigTree 601

Memory increases about 64MB per addBigTree call.

So both, performance and memory consumption are better by a factor
of 2!