Preliminary port of project file quick open data provider to Path.
While this shows some regressions in the benchmark (which can be
resolved later), the memory consumption is much better now since
we do not duplicate the string data of the file paths anymore.

Opening the Qt 4 source tree in KDevelop now only uses 134MB on my
machine, compared to 192MB with current master.