Fix color cache initialization from document on startup.
Apparently, when the ColorCache is setup, we do not yet have any
active document. Later on, when we get a documentActivated signal,
the document has no view yet but the text document only initializes
its text document upon creating the first view...

We can retry later on though which is dirty but seems to work for now.

Anyhow, this should probably be fixed such that a call to
TextDocument::textDocument() also forces the lazy initialization asap.