Revert "Never try to run a DVcsJob on a non-existing folder/file."
This reverts commit 6199cb30ebd6e1834fa621c97e56a9ed507845ac.

The reason is that the commit breaks the "Revision Graph..." context menu
entry and possibly others being invoked on files. Unlike what the API of
the class suggests, the git plugin happily supplies absolute filenames to
the constructor which hits the assert since the QDir::exists() check correctly
checks not only existence of the path, but also that its a directory.

Thats the downside of non-explicit default constructors like QDir(QString()).