Revert "Revert "Never try to run a DVcsJob on a non-existing folder/file.""
This reverts commit 98a6aafe4929e7f4b72df9a123ab4d38acdd2f14.

Turns out that the assert was justified, doing a revision graph on a
file simply hangs kdevelop. I first thought thats just because of the
repository size, but on a small repo it hangs too. So bring back the
assert which at least shows something is wrong in the code creating the
job.

This mess that some call git support desperately needs proper design.