Rudimentary port to the new IProjectManager interface using Path.
This is mostly done by replacing KUrl with Path until it compiles.
Especially the CMakeManager will require some more changes to get
it completely KUrl-free. The AbstractFileManagerPlugin-based project
managers do not need many changes at all.

We should really think about refactoring the CMakeManager such that it
is based on the AbstractFileManagerPlugin imo...