Optimize QuickOpen for project files using new Path.
FilterWithSeparator is now a PathFilter and uses the curiously
recurring template pattern instead of virtual inheritance to get the
Path for an element.

Especially the filtering is now much better as we do not have to
split the string item representation, since we can directly operate
on the existing path segments.