Fix activating documents with a non-empty range
This was broken with commit d77c94c1402916f8f00b02104f71187330548ecd which
replaced doc->setTextSelection with textView->setInitialRange in
documentcontroller.cpp without taking the extra logic from
TextDocument::setTextSelection into account. The setTextSelection function
not only sets the selection, but also sets the cursor which causes the
Kate view to scroll to the selected range. Without that the selection would
be set, but the view would not scroll leaving the user to think that the
activation didn't happen.

This affects at least the grepview plugin which tries to select the found
string in the occurrence that has been clicked upon.