Use a different color for informational items
Makes them distinguishable from warnings which is important since its a
different information level. I'm not 100% sure about using LinkText, the
other option would be ActiveText but its not really clear from the docs
which one would be better to use. The item is not really new or so, its
just clickable and has possibly important information.