Split up stopAction into a menu and an action
This allows to configure the shortcut for the stop-all-jobs action while
keeping the menu to stop individual jobs. Unfortunately it seems like
the KDE shortcut config dialog does not support shortcuts on KActionMenu.

Adjusted the texts where necessary, possibly a new icon for either of the
two would be good for the toolbar.

The main reason for making the shortcut configurable is that the current
default is somewhat annoying since many other things in the IDE use Escape
to exit some popup or so (find in kate for example) and thus it often
happens that one tabs Escape once too often and aborts a build etc.