Change default shortcut for stop-all-jobs
This makes the default shortcut a bit harder to hit and thus avoids hitting
it accidentally when trying to escape a kate passive popup or dialogs etc.

Unfortunately Ctrl+Escape is taken in KDE for the global shortcut of
ksysguardd, but this is now configurable so it can be easily changed for
users that don't have Ctrl+Escape taken already.