Fix filtered output item classification when multiple keywords are in the same line.
Currently, "warning" and "info"/"note" keywords, if found, override any
error keywords matched so far.

Fix this by scanning for all keywords in a consistent manner
(ERROR_INDICATORS => INDICATORS), and using the _first_ indicator in
line to classify the line.
This is somewhat empirical, but well: message type is usually put before
the message itself.

REVIEW: 107904