Second take on porting to the new IndexedString API.
This approach is much simpler and easier to review as it does not
try to refactor existing code at the same time. We will do that
later on, after we merged this branch.

What we did is basically getting rid of all usages of QByteArray
in the C++ parser and lexer. This also fixes parsing of code that
contains multibyte characters for free. Furthermore we can get
rid of tons of ugly conversions or even esoteric c-string comparisons.