Revert "Properly strip HTML from assistants' title and action descriptions."
This reverts commit b28f0aecc6347332d48f49f3b6271ee4a5b9c697.

The commit does not compile and cannot be made so, since it would
introduce a cyclic reference (interfaces cannot link against util). No
idea wether the new code in formattinghelpers should be moved into
interfaces or the whole implementation of iassistent moved elsewhere,
but those are the options I see at the moment.